feat: Implementar filtros por nivel de gobierno y corregir alcances en compromisos
- Backend: Filtrar normas por nivel de gobierno de la entidad del usuario
- Backend: Corregir compilación - mapear estado de CriterioEvaluacion (IdEstado)
- Backend: Actualizar CompromisoGobiernoDigital.Estado de int a FK
- Backend: Actualizar CreateCumplimientoCommand.CompromisoId de int a long
- Frontend: Corregir mapeo de alcances al editar compromisos (nombres a IDs)
- Frontend: Actualizar getEstados() para consumir API en lugar de valores hardcoded
- DB: Agregar scripts de migración para catálogos tipo_norma y estado_compromiso
- Docs: Agregar informes de pruebas, calidad y despliegue

+++ backend/PCM.Application/DTOs/CompromisoGobiernoDigital/CompromisoGobiernoDigitalDtos.cs
@@ -18,7 +18,7 @@ public class CreateCompromisoRequest
 
 public class UpdateCompromisoRequest
 {
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public string NombreCompromiso { get; set; } = string.Empty;
     public string? Descripcion { get; set; }
     public List<string> Alcances { get; set; } = new();
@@ -32,7 +32,7 @@ public class UpdateCompromisoRequest
 // Response DTOs
 public class CompromisoResponseDto
 {
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public string NombreCompromiso { get; set; } = string.Empty;
     public string? Descripcion { get; set; }
     public List<string> Alcances { get; set; } = new();
@@ -54,7 +54,7 @@ public class CompromisoNormativaDto
 public class CompromisoNormativaResponseDto
 {
     public int CompromisoNormativaId { get; set; }
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public int NormaId { get; set; }
     public string? NombreNorma { get; set; }
     public string? Numero { get; set; }
@@ -76,7 +76,7 @@ public class CriterioEvaluacionDto
 public class CriterioEvaluacionResponseDto
 {
     public int CriterioEvaluacionId { get; set; }
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public string Descripcion { get; set; } = string.Empty;
     public int Estado { get; set; } // FK a estado_compromiso
     public bool Activo { get; set; }
