+++ backend/PCM.Application/Features/Com1LiderGTD/Queries/GetCom1LiderGTDByEntidad/GetCom1LiderGTDByEntidadQuery.cs
@@ -6,6 +6,6 @@ namespace PCM.Application.Features.Com1LiderGTD.Queries.GetCom1LiderGTDByEntidad
 
 public class GetCom1LiderGTDByEntidadQuery : IRequest<Result<Com1LiderGTDResponse?>>
 {
-    public int CompromisoId { get; set; }
+    public long CompromisoId { get; set; }
     public Guid EntidadId { get; set; }
 }
