+++ backend/PCM.Domain/Entities/AlcanceCompromiso.cs
@@ -2,9 +2,9 @@ namespace PCM.Domain.Entities;
 
 public class AlcanceCompromiso
 {
-    public int AlcanceCompromisoId { get; set; }
-    public int CompromisoId { get; set; }
-    public int ClasificacionId { get; set; }
+    public long AlcanceCompromisoId { get; set; }
+    public long CompromisoId { get; set; }
+    public long ClasificacionId { get; set; }
     public bool Activo { get; set; } = true;
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
