+++ backend/PCM.Domain/Entities/Clasificacion.cs
@@ -5,7 +5,7 @@ namespace PCM.Domain.Entities;
 
 public class Clasificacion
 {
-    public int ClasificacionId { get; set; }
+    public long ClasificacionId { get; set; }
     public string Nombre { get; set; } = string.Empty;
     public string? Descripcion { get; set; }
     public bool Activo { get; set; } = true;
